SISTEMA BLAZE/meu-backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
alembic

python-jose[cryptography]
passlib[bcrypt]

pydantic
python-dotenv

requests
selenium
python-multipart
pytz


